Python 75 Commit

+++ photon_gen.c
@@ -76,7 +76,7 @@ int
 define_phot (p, f1, f2, nphot_tot, ioniz_or_final, iwind, freq_sampling)
      PhotPtr p;
      double f1, f2;
-     int nphot_tot;
+     double nphot_tot;
      int ioniz_or_final;
      int iwind;
      int freq_sampling;		// 0 --> old uniform approach, 1 --> mininimum fractions ins various bins
@@ -86,6 +86,7 @@ define_phot (p, f1, f2, nphot_tot, ioniz_or_final, iwind, freq_sampling)
   double ftot;
   int n;
   int iphot_start;
+
 
   if (freq_sampling == 0)
     {				/* Original approach, uniform sampling of entire wavelength interval,
